+++ settings/LoanPolicySettings.js
@@ -47,7 +47,7 @@ class LoanPolicySettings extends React.Component {
       type: 'okapi',
       records: 'fixedDueDateSchedules',
       path: 'fixed-due-date-schedule-storage/fixed-due-date-schedules',
-      dataKey: 'fixed-due-date-schedules',
+      resourceShouldRefresh: true
     },
   });